Rename oauth_open_id_connect_policy to openid_connect_policy
