+++ examples/data-sources/pingfederate_openid_connect_policy/data-source.tf
@@ -1,0 +1,3 @@
+data "pingfederate_openid_connect_policy" "myOIDCPolicy" {
+  policy_id = "myOIDCPolicyExampleId"
+}
