+++ examples/resources/pingfederate_openid_connect_policy/resource.tf
@@ -1,0 +1,190 @@
+resource "pingfederate_oauth_access_token_manager" "jwt_example" {
+  manager_id = "jsonWebTokenOatm"
+  name       = "jsonWebTokenExample"
+  plugin_descriptor_ref = {
+    id = "com.pingidentity.pf.access.token.management.plugins.JwtBearerAccessTokenManagementPlugin"
+  }
+  configuration = {
+    tables = [
+      {
+        name = "Symmetric Keys"
+        rows = [
+          {
+            fields = [
+              {
+                name  = "Key ID"
+                value = "keyidentifier"
+              },
+              {
+                name  = "Key"
+                value = var.access_token_manager_key
+              },
+              {
+                name  = "Encoding"
+                value = "b64u"
+              }
+            ]
+            default_row = false
+          }
+        ]
+      },
+      {
+        name = "Certificates"
+        rows = []
+      }
+    ]
+    fields = [
+      {
+        name  = "Token Lifetime"
+        value = "120"
+      },
+      {
+        name  = "Use Centralized Signing Key"
+        value = "false"
+      },
+      {
+        name  = "JWS Algorithm"
+        value = ""
+      },
+      {
+        name  = "Active Symmetric Key ID"
+        value = "keyidentifier"
+      },
+      {
+        name  = "Active Signing Certificate Key ID"
+        value = ""
+      },
+      {
+        name  = "JWE Algorithm"
+        value = "dir"
+      },
+      {
+        name  = "JWE Content Encryption Algorithm"
+        value = "A192CBC-HS384"
+      },
+      {
+        name  = "Active Symmetric Encryption Key ID"
+        value = "keyidentifier"
+      },
+      {
+        name  = "Asymmetric Encryption Key"
+        value = ""
+      },
+      {
+        name  = "Asymmetric Encryption JWKS URL"
+        value = ""
+      },
+      {
+        name  = "Enable Token Revocation"
+        value = "false"
+      },
+      {
+        name  = "Include Key ID Header Parameter"
+        value = "true"
+      },
+      {
+        name  = "Default JWKS URL Cache Duration"
+        value = "720"
+      },
+      {
+        name  = "Include JWE Key ID Header Parameter"
+        value = "true"
+      },
+      {
+        name  = "Client ID Claim Name"
+        value = "client_id"
+      },
+      {
+        name  = "Scope Claim Name"
+        value = "scope"
+      },
+      {
+        name  = "Space Delimit Scope Values"
+        value = "true"
+      },
+      {
+        name  = "Authorization Details Claim Name"
+        value = "authorization_details"
+      },
+      {
+        name  = "Issuer Claim Value"
+        value = ""
+      },
+      {
+        name  = "Audience Claim Value"
+        value = ""
+      },
+      {
+        name  = "JWT ID Claim Length"
+        value = "22"
+      },
+      {
+        name  = "Access Grant GUID Claim Name"
+        value = ""
+      },
+      {
+        name  = "JWKS Endpoint Path"
+        value = ""
+      },
+      {
+        name  = "JWKS Endpoint Cache Duration"
+        value = "720"
+      },
+      {
+        name  = "Expand Scope Groups"
+        value = "false"
+      },
+      {
+        name  = "Type Header Value"
+        value = ""
+      }
+    ]
+  }
+  attribute_contract = {
+    extended_attributes = [
+      {
+        name         = "contract"
+        multi_valued = false
+      }
+    ]
+  }
+  selection_settings = {
+    resource_uris = []
+  }
+  access_control_settings = {
+    restrict_clients = false
+  }
+  session_validation_settings = {
+    check_valid_authn_session       = false
+    check_session_revocation_status = false
+    update_authn_session_activity   = false
+    include_session_id              = false
+  }
+}
+
+resource "pingfederate_openid_connect_policy" "OIDCPolicy" {
+  policy_id = "oidcPolicy"
+  name      = "oidcPolicy"
+  access_token_manager_ref = {
+    id = pingfederate_oauth_access_token_manager.jwt_example.manager_id
+  }
+  attribute_contract = {
+    extended_attributes = []
+  }
+  attribute_mapping = {
+    attribute_contract_fulfillment = {
+      "sub" = {
+        source = {
+          type = "TOKEN"
+        }
+        value = "Username"
+      }
+    }
+  }
+  return_id_token_on_refresh_grant = false
+  include_sri_in_id_token          = false
+  include_s_hash_in_id_token       = false
+  include_user_info_in_id_token    = false
+  reissue_id_token_in_hybrid_flow  = false
+  id_token_lifetime                = 5
+}
