+++ internal/provider/provider.go
@@ -702,6 +702,7 @@ func (p *pingfederateProvider) DataSources(_ context.Context) []func() datasourc
 		oauthclient.OauthClientDataSource,
 		oauthissuer.OauthIssuerDataSource,
 		oauthtokenexchangetokengeneratormapping.OauthTokenExchangeTokenGeneratorMappingDataSource,
+		oauthopenidconnectpolicy.OpenidConnectPolicyDataSource,
 		oauthopenidconnectpolicy.OauthOpenIdConnectPolicyDataSource,
 		passwordcredentialvalidator.PasswordCredentialValidatorDataSource,
 		protocolmetadatalifetimesettings.ProtocolMetadataLifetimeSettingsDataSource,
@@ -773,6 +774,7 @@ func (p *pingfederateProvider) Resources(_ context.Context) []func() resource.Re
 		oauthclient.OauthClientResource,
 		oauthidpadaptermappings.OauthIdpAdapterMappingResource,
 		oauthissuer.OauthIssuerResource,
+		oauthopenidconnectpolicy.OpenidConnectPolicyResource,
 		oauthopenidconnectpolicy.OauthOpenIdConnectPolicyResource,
 		oauthopenidconnectsettings.DeprecatedOidcResource,
 		oauthopenidconnectsettings.OpenidConnectSettingsResource,
