+++ internal/resource/config/oauth/openidconnect/policy/openid_connect_policy_common.go
